Moved method searchDepartments to DepartmentService
