+++ ejbModule/de/hallerweb/enterprise/prioritize/view/boundary/CompanyService.java
@@ -17,13 +17,9 @@
 package de.hallerweb.enterprise.prioritize.view.boundary;
 
 import de.hallerweb.enterprise.prioritize.controller.CompanyController;
-import de.hallerweb.enterprise.prioritize.controller.search.SearchController;
 import de.hallerweb.enterprise.prioritize.controller.security.AuthorizationController;
 import de.hallerweb.enterprise.prioritize.controller.security.RestAccessController;
-import de.hallerweb.enterprise.prioritize.controller.security.UserRoleController;
 import de.hallerweb.enterprise.prioritize.model.Company;
-import de.hallerweb.enterprise.prioritize.model.Department;
-import de.hallerweb.enterprise.prioritize.model.search.SearchResult;
 import de.hallerweb.enterprise.prioritize.model.security.User;
 import jakarta.ejb.EJB;
 import jakarta.enterprise.context.RequestScoped;
@@ -33,17 +29,14 @@ import jakarta.ws.rs.core.Response;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 /**
- *
  * <p>
  * Copyright: (c) 2015
  * </p>
  *
- *  <p>
+ * <p>
  * Peter Haller
  * </p>
  *
@@ -58,10 +51,6 @@ public class CompanyService {
 
     @EJB
     CompanyController companyController;
-    @EJB
-    UserRoleController userRoleController;
-    @EJB
-    SearchController searchController;
     @EJB
     RestAccessController accessController;
     @EJB
@@ -127,54 +116,4 @@ public class CompanyService {
         }
     }
 
-
-    /**
-     * Returns all the departments matching the seach phrase.
-     *
-     * @api {get} /search/departments searchDepartments
-     * @apiName searchDepartments
-     * @apiGroup /company
-     * @apiDescription Searches all departments which contain the given phrase
-     * @apiParam {String} apiKey The API-Key of the user accessing the service.
-     * @apiParam {String} phrase The search phrase used in the search.
-     * @apiSuccess {Department} department JSON department Objects which contained the search phrase.
-     * @apiSuccessExample Success-Response:
-     * HTTP/1.1 200 OK
-     * [
-     * {
-     * "id" : 6,
-     * "address" : {
-     * "id" : 5,
-     * "zipCode" : "00000",
-     * "phone" : "00000-00000",
-     * "fax" : "00000-00000",
-     * "city" : "City of Admin",
-     * "street" : "Street of Admins"
-     * ...many more
-     * }
-     * ]
-     * @apiError NotAuthorized  APIKey incorrect.
-     */
-    @GET
-    @Path("search/departments")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Set<Department> searchDepartments(@QueryParam(value = "apiKey") String apiKey, @QueryParam(value = "phrase") String phrase) {
-        User user = accessController.checkApiKey(apiKey);
-        if (user != null) {
-            Set<Department> searchResult = new HashSet<>();
-            List<SearchResult> results = searchController.searchDepartments(phrase, user);
-            for (SearchResult result : results) {
-                Department dept = (Department) result.getResult();
-                if (authController.canRead(dept, user)) {
-                    searchResult.add(dept);
-                } else {
-                    throw new NotAuthorizedException(Response.serverError());
-                }
-            }
-            return searchResult;
-
-        } else {
-            throw new NotAuthorizedException(Response.serverError());
-        }
-    }
 }
